move to tag package as it's not internal
